loader: Move load_dtb to top-level module
This method can also be used when loading ELF files, hence move it to
the loader module level.
